moved patient data to use context
