+++ src/Utils/Web5Provider.jsx
@@ -1,40 +1,37 @@
-import { createContext, useEffect, useState } from 'react';
-import { Web5 } from '@web5/api/browser';
-import propTypes from 'prop-types';
-import Swal from 'sweetalert2';
+import { createContext, useEffect, useState } from "react";
+import { Web5 } from "@web5/api/browser";
+import propTypes from "prop-types";
+import Swal from "sweetalert2";
 
 export const Web5Context = createContext(null);
 
 export const Web5Provider = ({ children }) => {
-    Web5Provider.propTypes = {
-        children: propTypes.string,
-    };
+  Web5Provider.propTypes = {
+    children: propTypes.string,
+  };
 
   const [web5, setWeb5] = useState(null);
   const [myDID, setMyDID] = useState(null);
+  const [patientData, setPatientData] = useState([]);
 
   useEffect(() => {
-
-   
-
     const connectToWeb5 = async () => {
       try {
         const { web5: connectedWeb5, did: connectedDID } = await Web5.connect();
         setWeb5(connectedWeb5);
         setMyDID(connectedDID);
-       
+
         if (connectedWeb5 && connectedDID) {
           await configureProtocol(connectedWeb5, connectedDID);
-        }else{
-          console.log('hey');
+        } else {
+          console.log("hey");
         }
       } catch (error) {
-        console.log('error', error);
-        Swal.fire({  
-            icon: 'error',
-            text: 'Failed to connect to Web5',
+        console.log("error", error);
+        Swal.fire({
+          icon: "error",
+          text: "Failed to connect to Web5",
         });
-        
       }
     };
 
@@ -43,25 +40,64 @@ export const Web5Provider = ({ children }) => {
     const configureProtocol = async (web5, myDID) => {
       const protocolDefinition = newProtocolDefinition();
       const protocolUrl = protocolDefinition.protocol;
-  
+
       // const { protocols: localProtocols, status: localProtocolStatus } = await queryLocalProtocol(web5, protocolUrl);
       // if (localProtocolStatus.code !== 200 || localProtocols.length === 0) {
       //   const result = await installLocalProtocol(web5, protocolDefinition);
       //   console.log({ result })
       //   console.log("Protocol installed locally");
       // }
-  
-      const { protocols: remoteProtocols, status: remoteProtocolStatus } = await queryRemoteProtocol(web5, myDID, protocolUrl);
+
+      const { protocols: remoteProtocols, status: remoteProtocolStatus } =
+        await queryRemoteProtocol(web5, myDID, protocolUrl);
       if (remoteProtocolStatus.code !== 200 || remoteProtocols.length === 0) {
-        const result = await installRemoteProtocol(web5, myDID, protocolDefinition);
-        console.log({ result })
+        const result = await installRemoteProtocol(
+          web5,
+          myDID,
+          protocolDefinition
+        );
+        console.log({ result });
         console.log("Protocol installed remotely");
       }
     };
-    
-  }, []);
 
-  
+    readAllDataFromDWNForDid();
+  }, [myDID]);
+
+  const readAllDataFromDWNForDid = async () => {
+    const response = await web5.dwn.records.query({
+      from: myDID,
+      message: {
+        filter: {
+          schema: "https://med-5.vercel.app/schema/patientRecord",
+          dataFormat: "application/json",
+        },
+      },
+    });
+
+    let num = 1;
+
+    let arr = [];
+
+    for (const record of response.records) {
+      let { record: readRecord } = await web5.dwn.records.read({
+        message: {
+          filter: {
+            recordId: record.id,
+          },
+        },
+      });
+
+      // assuming the record has a text payload
+      const text = await readRecord.data.text();
+
+      arr.push(JSON.parse(text));
+
+      num += 1;
+    }
+
+    setPatientData(arr);
+  };
 
   const newProtocolDefinition = () => {
     return {
@@ -72,7 +108,6 @@ export const Web5Provider = ({ children }) => {
           schema: "https://med-5.vercel.app/schema/patientRecord",
           dataFormats: ["application/json"],
         },
-       
       },
       structure: {
         patientRecord: {
@@ -96,7 +131,6 @@ export const Web5Provider = ({ children }) => {
   //     },
   //   });
   // };
-
 
   const queryRemoteProtocol = async (web5, myDID) => {
     return await web5.dwn.protocols.query({
@@ -123,7 +157,7 @@ export const Web5Provider = ({ children }) => {
         definition: protocolDefinition,
       },
     });
-    console.log('Remote protocol installed');
+    console.log("Remote protocol installed");
     return await protocol.send(myDID);
   };
 
@@ -137,4 +171,3 @@ export const Web5Provider = ({ children }) => {
     </>
   );
 };
-
